fix: rewrite inclusion of the feature in MainActivity thereby fixing feature
