+++ app/src/main/java/org/schabi/newpipe/MainActivity.java
@@ -20,37 +20,51 @@
 
 package org.schabi.newpipe;
 
-import static org.schabi.newpipe.util.Localization.assureCorrectAppLanguage;
-
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
+import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.os.Build;
 import android.os.Bundle;
+import android.os.Handler;
+import android.os.Looper;
 import android.util.Log;
 import android.view.KeyEvent;
+import android.view.LayoutInflater;
 import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
+import android.view.ViewGroup;
+import android.widget.AdapterView;
+import android.widget.ArrayAdapter;
 import android.widget.FrameLayout;
+import android.widget.Spinner;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.ActionBarDrawerToggle;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
 import androidx.core.view.GravityCompat;
 import androidx.drawerlayout.widget.DrawerLayout;
 import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.preference.PreferenceManager;
 
 import com.google.android.material.bottomsheet.BottomSheetBehavior;
 
 import org.schabi.newpipe.databinding.ActivityMainBinding;
 import org.schabi.newpipe.databinding.DrawerHeaderBinding;
 import org.schabi.newpipe.databinding.DrawerLayoutBinding;
+import org.schabi.newpipe.databinding.InstanceSpinnerLayoutBinding;
 import org.schabi.newpipe.databinding.ToolbarLayoutBinding;
 import org.schabi.newpipe.error.ErrorActivity;
+import org.schabi.newpipe.extractor.NewPipe;
 import org.schabi.newpipe.extractor.StreamingService;
-import org.schabi.newpipe.fragments.DrawerFragment;
+import org.schabi.newpipe.extractor.exceptions.ExtractionException;
+import org.schabi.newpipe.extractor.services.peertube.PeertubeInstance;
 import org.schabi.newpipe.fragments.BackPressable;
 import org.schabi.newpipe.fragments.MainFragment;
 import org.schabi.newpipe.fragments.detail.VideoDetailFragment;
@@ -59,18 +73,27 @@ import org.schabi.newpipe.player.Player;
 import org.schabi.newpipe.player.event.OnKeyDownListener;
 import org.schabi.newpipe.player.helper.PlayerHolder;
 import org.schabi.newpipe.player.playqueue.PlayQueue;
+import org.schabi.newpipe.settings.sections.Section;
+import org.schabi.newpipe.settings.sections.SectionsManager;
 import org.schabi.newpipe.util.Constants;
 import org.schabi.newpipe.util.DeviceUtils;
+import org.schabi.newpipe.util.KioskTranslator;
 import org.schabi.newpipe.util.Localization;
 import org.schabi.newpipe.util.NavigationHelper;
+import org.schabi.newpipe.util.PeertubeHelper;
 import org.schabi.newpipe.util.PermissionHelper;
 import org.schabi.newpipe.util.SerializedCache;
 import org.schabi.newpipe.util.ServiceHelper;
 import org.schabi.newpipe.util.StateSaver;
 import org.schabi.newpipe.util.TLSSocketFactoryCompat;
 import org.schabi.newpipe.util.ThemeHelper;
+import org.schabi.newpipe.views.FocusOverlayView;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
+
+import static org.schabi.newpipe.util.Localization.assureCorrectAppLanguage;
 
 public class MainActivity extends AppCompatActivity {
     private static final String TAG = "MainActivity";
@@ -82,11 +105,29 @@ public class MainActivity extends AppCompatActivity {
     private DrawerLayoutBinding drawerLayoutBinding;
     private ToolbarLayoutBinding toolbarLayoutBinding;
 
+    private SectionsManager sectionsManager;
+    private List<Section> sectionList = new ArrayList<>();
+
+    private ActionBarDrawerToggle toggle;
+
     private boolean servicesShown = false;
 
-    DrawerFragment drawer;
-    DrawerLayout drawerLayout;
     private BroadcastReceiver broadcastReceiver;
+
+    static final int ITEM_ID_BLANK = -1;
+    static final int ITEM_ID_SETTINGS = -2;
+    static final int ITEM_ID_ABOUT = -3;
+    static final int ITEM_ID_BOOKMARKS = -4;
+    static final int ITEM_ID_FEED = -5;
+    static final int ITEM_ID_SUBSCRIPTIONS = -6;
+    static final int ITEM_ID_DOWNLOADS = -7;
+    static final int ITEM_ID_HISTORY = -8;
+    static final int ITEM_ID_DEFAULT_KIOSK = -9;
+    static final int ITEM_ID_KIOSK = -10;
+    static final int ITEM_ID_CHANNEL = -11;
+    static final int ITEM_ID_PLAYLIST = -12;
+
+    private static final int ORDER = 0;
 
     /*//////////////////////////////////////////////////////////////////////////
     // Activity's LifeCycle
@@ -106,8 +147,8 @@ public class MainActivity extends AppCompatActivity {
 
         ThemeHelper.setDayNightMode(this);
         ThemeHelper.setTheme(this, ServiceHelper.getSelectedServiceId(this));
-        setContentView(R.layout.activity_main);
-        setSupportActionBar(findViewById(R.id.toolbar));
+
+        assureCorrectAppLanguage(this);
         super.onCreate(savedInstanceState);
 
         mainBinding = ActivityMainBinding.inflate(getLayoutInflater());
@@ -120,6 +161,386 @@ public class MainActivity extends AppCompatActivity {
         if (getSupportFragmentManager().getBackStackEntryCount() == 0) {
             initFragments();
         }
+
+        setSupportActionBar(toolbarLayoutBinding.toolbar);
+        try {
+            setupDrawer();
+        } catch (final Exception e) {
+            ErrorActivity.reportUiErrorInSnackbar(this, "Setting up drawer", e);
+        }
+
+        if (DeviceUtils.isTv(this)) {
+            FocusOverlayView.setupFocusObserver(this);
+        }
+        openMiniPlayerUponPlayerStarted();
+    }
+
+    private void setupDrawer() throws Exception {
+        sectionsManager = SectionsManager.getManager(this);
+        sectionsManager.setSavedSectionsListener(() -> {
+            setupSections();
+        });
+
+
+//        //Tabs
+//        final int currentServiceId = ServiceHelper.getSelectedServiceId(this);
+//        final StreamingService service = NewPipe.getService(currentServiceId);
+//
+//        int kioskId = 0;
+//
+//        for (final String ks : service.getKioskList().getAvailableKiosks()) {
+//            drawerLayoutBinding.navigation.getMenu()
+//                    .add(R.id.menu_tabs_group, kioskId, 0, KioskTranslator
+//                            .getTranslatedKioskName(ks, this))
+//                    .setIcon(KioskTranslator.getKioskIcon(ks, this));
+//            kioskId++;
+//        }
+//
+//        drawerLayoutBinding.navigation.getMenu()
+//                .add(R.id.menu_tabs_group, ITEM_ID_SUBSCRIPTIONS, ORDER,
+//                        R.string.tab_subscriptions)
+//                .setIcon(R.drawable.ic_tv);
+//        drawerLayoutBinding.navigation.getMenu()
+//                .add(R.id.menu_tabs_group, ITEM_ID_FEED, ORDER, R.string.fragment_feed_title)
+//                .setIcon(R.drawable.ic_rss_feed);
+//        drawerLayoutBinding.navigation.getMenu()
+//                .add(R.id.menu_tabs_group, ITEM_ID_BOOKMARKS, ORDER, R.string.tab_bookmarks)
+//                .setIcon(R.drawable.ic_bookmark);
+//        drawerLayoutBinding.navigation.getMenu()
+//                .add(R.id.menu_tabs_group, ITEM_ID_DOWNLOADS, ORDER, R.string.downloads)
+//                .setIcon(R.drawable.ic_file_download);
+//        drawerLayoutBinding.navigation.getMenu()
+//                .add(R.id.menu_tabs_group, ITEM_ID_HISTORY, ORDER, R.string.action_history)
+//                .setIcon(R.drawable.ic_history);
+//
+//        //Settings and About
+//        drawerLayoutBinding.navigation.getMenu()
+//                .add(R.id.menu_options_about_group, ITEM_ID_SETTINGS, ORDER, R.string.settings)
+//                .setIcon(R.drawable.ic_settings);
+//        drawerLayoutBinding.navigation.getMenu()
+//                .add(R.id.menu_options_about_group, ITEM_ID_ABOUT, ORDER, R.string.tab_about)
+//                .setIcon(R.drawable.ic_info_outline);
+//
+        setupSections();
+        toggle = new ActionBarDrawerToggle(this, mainBinding.getRoot(),
+                toolbarLayoutBinding.toolbar, R.string.drawer_open, R.string.drawer_close);
+        toggle.syncState();
+        mainBinding.getRoot().addDrawerListener(toggle);
+        mainBinding.getRoot().addDrawerListener(new DrawerLayout.SimpleDrawerListener() {
+            private int lastService;
+
+            @Override
+            public void onDrawerOpened(final View drawerView) {
+                lastService = ServiceHelper.getSelectedServiceId(MainActivity.this);
+            }
+
+            @Override
+            public void onDrawerClosed(final View drawerView) {
+                if (servicesShown) {
+                    toggleServices();
+                }
+                if (lastService != ServiceHelper.getSelectedServiceId(MainActivity.this)) {
+                    ActivityCompat.recreate(MainActivity.this);
+                }
+            }
+        });
+
+
+        drawerLayoutBinding.navigation.setNavigationItemSelectedListener(this::drawerItemSelected);
+        setupDrawerHeader();
+    }
+
+
+    public void setupSections() {
+        sectionList.clear();
+        sectionList.addAll(sectionsManager.getSections());
+
+        updateSectionsIconAndDescription();
+    }
+
+    private void updateSectionsIconAndDescription() {
+        drawerLayoutBinding.navigation.getMenu().clear();
+
+        int kioskCounter = 0;
+        final List<String> kioskList = getKioskIdsAsList();
+        final int serviceId = ServiceHelper.getSelectedServiceId(this);
+
+        for (int i = 0; i < sectionList.size(); i++) {
+            final Section section = sectionList.get(i);
+
+            switch (section.getSectionId()) {
+                case ITEM_ID_BLANK:
+                    //don't add blank sections
+                    break;
+                case ITEM_ID_DEFAULT_KIOSK:
+                case ITEM_ID_KIOSK:
+                    //limits number of Kiosks to numbers of actual Kiosks
+                    if (kioskCounter < kioskList.size()) {
+                        final String kioskId = kioskList.get(kioskCounter);
+                        final String sectionName =
+                                KioskTranslator.getTranslatedKioskName(kioskId, this);
+                        final int iconID = KioskTranslator.getKioskIcon(kioskId, this);
+                        drawerLayoutBinding.navigation.getMenu().add(R.id.menu_tabs_group,
+                                serviceId, ORDER,
+                                sectionName)
+                                .setIcon(iconID);
+                        kioskCounter++;
+                    }
+                    break;
+                default:
+                    drawerLayoutBinding.navigation.getMenu()
+                            .add(R.id.menu_tabs_group, section.getSectionId(), ORDER,
+                                    section.getSectionName(this))
+                            .setIcon(section.getSectionIconRes(this));
+            }
+        }
+
+
+        //Settings and About
+        drawerLayoutBinding.navigation.getMenu()
+                .add(R.id.menu_options_about_group, ITEM_ID_SETTINGS, ORDER, R.string.settings)
+                .setIcon(R.drawable.ic_settings);
+        drawerLayoutBinding.navigation.getMenu()
+                .add(R.id.menu_options_about_group, ITEM_ID_ABOUT, ORDER, R.string.tab_about)
+                .setIcon(R.drawable.ic_info_outline);
+    }
+
+    private List<String> getKioskIdsAsList() {
+        final int serviceId = ServiceHelper.getSelectedServiceId(this);
+        final StreamingService service;
+        final List<String> kioskList = new ArrayList<>();
+        try {
+            service = NewPipe.getService(serviceId);
+            kioskList.addAll(service.getKioskList().getAvailableKiosks());
+        } catch (final ExtractionException e) {
+            e.printStackTrace();
+        }
+        return kioskList;
+    }
+
+
+    private boolean drawerItemSelected(final MenuItem item) {
+        switch (item.getGroupId()) {
+            case R.id.menu_services_group:
+                changeService(item);
+                break;
+            case R.id.menu_tabs_group:
+                try {
+                    tabSelected(item);
+                } catch (final Exception e) {
+                    ErrorActivity.reportUiErrorInSnackbar(this, "Selecting main page tab", e);
+                }
+                break;
+            case R.id.menu_options_about_group:
+                optionsAboutSelected(item);
+                break;
+            default:
+                return false;
+        }
+
+        mainBinding.getRoot().closeDrawers();
+        return true;
+    }
+
+    private void changeService(final MenuItem item) {
+        drawerLayoutBinding.navigation.getMenu()
+                .getItem(ServiceHelper.getSelectedServiceId(this))
+                .setChecked(false);
+        ServiceHelper.setSelectedServiceId(this, item.getItemId());
+        drawerLayoutBinding.navigation.getMenu()
+                .getItem(ServiceHelper.getSelectedServiceId(this))
+                .setChecked(true);
+    }
+
+    private void tabSelected(final MenuItem item) throws ExtractionException {
+        switch (item.getItemId()) {
+            case ITEM_ID_SUBSCRIPTIONS:
+                NavigationHelper.openSubscriptionFragment(getSupportFragmentManager());
+                break;
+            case ITEM_ID_FEED:
+                NavigationHelper.openFeedFragment(getSupportFragmentManager());
+                break;
+            case ITEM_ID_BOOKMARKS:
+                NavigationHelper.openBookmarksFragment(getSupportFragmentManager());
+                break;
+            case ITEM_ID_DOWNLOADS:
+                NavigationHelper.openDownloads(this);
+                break;
+            case ITEM_ID_HISTORY:
+                NavigationHelper.openStatisticFragment(getSupportFragmentManager());
+                break;
+            default:
+                final int currentServiceId = ServiceHelper.getSelectedServiceId(this);
+                final StreamingService service = NewPipe.getService(currentServiceId);
+                String serviceName = "";
+
+                int kioskId = 0;
+                for (final String ks : service.getKioskList().getAvailableKiosks()) {
+                    if (kioskId == item.getItemId()) {
+                        serviceName = ks;
+                    }
+                    kioskId++;
+                }
+
+                NavigationHelper.openKioskFragment(getSupportFragmentManager(), currentServiceId,
+                        serviceName);
+                break;
+        }
+    }
+
+    private void optionsAboutSelected(final MenuItem item) {
+        switch (item.getItemId()) {
+            case ITEM_ID_SETTINGS:
+                NavigationHelper.openSettings(this);
+                break;
+            case ITEM_ID_ABOUT:
+                NavigationHelper.openAbout(this);
+                break;
+        }
+    }
+
+    private void setupDrawerHeader() {
+        drawerHeaderBinding.drawerHeaderActionButton.setOnClickListener(view -> toggleServices());
+
+        // If the current app name is bigger than the default "NewPipe" (7 chars),
+        // let the text view grow a little more as well.
+        if (getString(R.string.app_name).length() > "NewPipe".length()) {
+            final ViewGroup.LayoutParams layoutParams =
+                    drawerHeaderBinding.drawerHeaderNewpipeTitle.getLayoutParams();
+            layoutParams.width = ViewGroup.LayoutParams.WRAP_CONTENT;
+            drawerHeaderBinding.drawerHeaderNewpipeTitle.setLayoutParams(layoutParams);
+            drawerHeaderBinding.drawerHeaderNewpipeTitle.setMaxLines(2);
+            drawerHeaderBinding.drawerHeaderNewpipeTitle.setMinWidth(getResources()
+                    .getDimensionPixelSize(R.dimen.drawer_header_newpipe_title_default_width));
+            drawerHeaderBinding.drawerHeaderNewpipeTitle.setMaxWidth(getResources()
+                    .getDimensionPixelSize(R.dimen.drawer_header_newpipe_title_max_width));
+        }
+    }
+
+    private void toggleServices() {
+        servicesShown = !servicesShown;
+
+        drawerLayoutBinding.navigation.getMenu().removeGroup(R.id.menu_services_group);
+        drawerLayoutBinding.navigation.getMenu().removeGroup(R.id.menu_tabs_group);
+        drawerLayoutBinding.navigation.getMenu().removeGroup(R.id.menu_options_about_group);
+
+        if (servicesShown) {
+            showServices();
+        } else {
+            try {
+                showTabs();
+            } catch (final Exception e) {
+                ErrorActivity.reportUiErrorInSnackbar(this, "Showing main page tabs", e);
+            }
+        }
+    }
+
+    private void showServices() {
+        drawerHeaderBinding.drawerArrow.setImageResource(R.drawable.ic_arrow_drop_up);
+
+        for (final StreamingService s : NewPipe.getServices()) {
+            final String title = s.getServiceInfo().getName()
+                    + (ServiceHelper.isBeta(s) ? " (beta)" : "");
+
+            final MenuItem menuItem = drawerLayoutBinding.navigation.getMenu()
+                    .add(R.id.menu_services_group, s.getServiceId(), ORDER, title)
+                    .setIcon(ServiceHelper.getIcon(s.getServiceId()));
+
+            // peertube specifics
+            if (s.getServiceId() == 3) {
+                enhancePeertubeMenu(s, menuItem);
+            }
+        }
+        drawerLayoutBinding.navigation.getMenu()
+                .getItem(ServiceHelper.getSelectedServiceId(this))
+                .setChecked(true);
+    }
+
+    private void enhancePeertubeMenu(final StreamingService s, final MenuItem menuItem) {
+        final PeertubeInstance currentInstance = PeertubeHelper.getCurrentInstance();
+        menuItem.setTitle(currentInstance.getName() + (ServiceHelper.isBeta(s) ? " (beta)" : ""));
+        final Spinner spinner = InstanceSpinnerLayoutBinding.inflate(LayoutInflater.from(this))
+                .getRoot();
+        final List<PeertubeInstance> instances = PeertubeHelper.getInstanceList(this);
+        final List<String> items = new ArrayList<>();
+        int defaultSelect = 0;
+        for (final PeertubeInstance instance : instances) {
+            items.add(instance.getName());
+            if (instance.getUrl().equals(currentInstance.getUrl())) {
+                defaultSelect = items.size() - 1;
+            }
+        }
+        final ArrayAdapter<String> adapter = new ArrayAdapter<>(this,
+                R.layout.instance_spinner_item, items);
+        adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
+        spinner.setAdapter(adapter);
+        spinner.setSelection(defaultSelect, false);
+        spinner.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
+            @Override
+            public void onItemSelected(final AdapterView<?> parent, final View view,
+                                       final int position, final long id) {
+                final PeertubeInstance newInstance = instances.get(position);
+                if (newInstance.getUrl().equals(PeertubeHelper.getCurrentInstance().getUrl())) {
+                    return;
+                }
+                PeertubeHelper.selectInstance(newInstance, getApplicationContext());
+                changeService(menuItem);
+                mainBinding.getRoot().closeDrawers();
+                new Handler(Looper.getMainLooper()).postDelayed(() -> {
+                    getSupportFragmentManager().popBackStack(null,
+                            FragmentManager.POP_BACK_STACK_INCLUSIVE);
+                    ActivityCompat.recreate(MainActivity.this);
+                }, 300);
+            }
+
+            @Override
+            public void onNothingSelected(final AdapterView<?> parent) {
+
+            }
+        });
+        menuItem.setActionView(spinner);
+    }
+
+    private void showTabs() throws ExtractionException {
+        drawerHeaderBinding.drawerArrow.setImageResource(R.drawable.ic_arrow_drop_down);
+
+        //Tabs
+        final int currentServiceId = ServiceHelper.getSelectedServiceId(this);
+        final StreamingService service = NewPipe.getService(currentServiceId);
+
+        int kioskId = 0;
+
+        for (final String ks : service.getKioskList().getAvailableKiosks()) {
+            drawerLayoutBinding.navigation.getMenu()
+                    .add(R.id.menu_tabs_group, kioskId, ORDER,
+                            KioskTranslator.getTranslatedKioskName(ks, this))
+                    .setIcon(KioskTranslator.getKioskIcon(ks, this));
+            kioskId++;
+        }
+
+        drawerLayoutBinding.navigation.getMenu()
+                .add(R.id.menu_tabs_group, ITEM_ID_SUBSCRIPTIONS, ORDER, R.string.tab_subscriptions)
+                .setIcon(R.drawable.ic_tv);
+        drawerLayoutBinding.navigation.getMenu()
+                .add(R.id.menu_tabs_group, ITEM_ID_FEED, ORDER, R.string.fragment_feed_title)
+                .setIcon(R.drawable.ic_rss_feed);
+        drawerLayoutBinding.navigation.getMenu()
+                .add(R.id.menu_tabs_group, ITEM_ID_BOOKMARKS, ORDER, R.string.tab_bookmarks)
+                .setIcon(R.drawable.ic_bookmark);
+        drawerLayoutBinding.navigation.getMenu()
+                .add(R.id.menu_tabs_group, ITEM_ID_DOWNLOADS, ORDER, R.string.downloads)
+                .setIcon(R.drawable.ic_file_download);
+        drawerLayoutBinding.navigation.getMenu()
+                .add(R.id.menu_tabs_group, ITEM_ID_HISTORY, ORDER, R.string.action_history)
+                .setIcon(R.drawable.ic_history);
+
+        //Settings and About
+        drawerLayoutBinding.navigation.getMenu()
+                .add(R.id.menu_options_about_group, ITEM_ID_SETTINGS, ORDER, R.string.settings)
+                .setIcon(R.drawable.ic_settings);
+        drawerLayoutBinding.navigation.getMenu()
+                .add(R.id.menu_options_about_group, ITEM_ID_ABOUT, ORDER, R.string.tab_about)
+                .setIcon(R.drawable.ic_info_outline);
     }
 
     @Override
@@ -139,6 +560,48 @@ public class MainActivity extends AppCompatActivity {
         // Change the date format to match the selected language on resume
         Localization.initPrettyTime(Localization.resolvePrettyTime(getApplicationContext()));
         super.onResume();
+
+        // Close drawer on return, and don't show animation,
+        // so it looks like the drawer isn't open when the user returns to MainActivity
+        mainBinding.getRoot().closeDrawer(GravityCompat.START, false);
+        try {
+            final int selectedServiceId = ServiceHelper.getSelectedServiceId(this);
+            final String selectedServiceName = NewPipe.getService(selectedServiceId)
+                    .getServiceInfo().getName();
+            drawerHeaderBinding.drawerHeaderServiceView.setText(selectedServiceName);
+            drawerHeaderBinding.drawerHeaderServiceIcon.setImageResource(ServiceHelper
+                    .getIcon(selectedServiceId));
+
+            drawerHeaderBinding.drawerHeaderServiceView.post(() -> drawerHeaderBinding
+                    .drawerHeaderServiceView.setSelected(true));
+            drawerHeaderBinding.drawerHeaderActionButton.setContentDescription(
+                    getString(R.string.drawer_header_description) + selectedServiceName);
+        } catch (final Exception e) {
+            ErrorActivity.reportUiErrorInSnackbar(this, "Setting up service toggle", e);
+        }
+
+        final SharedPreferences sharedPreferences
+                = PreferenceManager.getDefaultSharedPreferences(this);
+        if (sharedPreferences.getBoolean(Constants.KEY_THEME_CHANGE, false)) {
+            if (DEBUG) {
+                Log.d(TAG, "Theme has changed, recreating activity...");
+            }
+            sharedPreferences.edit().putBoolean(Constants.KEY_THEME_CHANGE, false).apply();
+            ActivityCompat.recreate(this);
+        }
+
+        if (sharedPreferences.getBoolean(Constants.KEY_MAIN_PAGE_CHANGE, false)) {
+            if (DEBUG) {
+                Log.d(TAG, "main page has changed, recreating main fragment...");
+            }
+            sharedPreferences.edit().putBoolean(Constants.KEY_MAIN_PAGE_CHANGE, false).apply();
+            NavigationHelper.openMainActivity(this);
+        }
+
+        final boolean isHistoryEnabled = sharedPreferences.getBoolean(
+                getString(R.string.enable_watch_history_key), true);
+        drawerLayoutBinding.navigation.getMenu().findItem(ITEM_ID_HISTORY)
+                .setVisible(isHistoryEnabled);
     }
 
     @Override
@@ -247,29 +710,6 @@ public class MainActivity extends AppCompatActivity {
         }
     }
 
-    @Override
-    public boolean onCreateOptionsMenu(final Menu menu) {
-        if (DEBUG) {
-            Log.d(TAG, "onCreateOptionsMenu() called with: menu = [" + menu + "]");
-        }
-
-        final Fragment fragment = getSupportFragmentManager().
-                findFragmentById(R.id.fragment_holder);
-        if (!(fragment instanceof SearchFragment)) {
-            findViewById(R.id.toolbar).findViewById(R.id.toolbar_search_container)
-                    .setVisibility(View.GONE);
-        }
-
-        updateDrawerNavigation();
-
-        return true;
-    }
-
-    public void setDrawer(final DrawerFragment drawer) {
-        this.drawer = drawer;
-        this.drawerLayout = drawer.getDrawer();
-    }
-
     /**
      * Implement the following diagram behavior for the up button:
      * <pre>
@@ -302,6 +742,45 @@ public class MainActivity extends AppCompatActivity {
         }
     }
 
+    /*//////////////////////////////////////////////////////////////////////////
+    // Menu
+    //////////////////////////////////////////////////////////////////////////*/
+
+    @Override
+    public boolean onCreateOptionsMenu(final Menu menu) {
+        if (DEBUG) {
+            Log.d(TAG, "onCreateOptionsMenu() called with: menu = [" + menu + "]");
+        }
+        super.onCreateOptionsMenu(menu);
+
+        final Fragment fragment
+                = getSupportFragmentManager().findFragmentById(R.id.fragment_holder);
+        if (!(fragment instanceof SearchFragment)) {
+            toolbarLayoutBinding.toolbarSearchContainer.getRoot().setVisibility(View.GONE);
+        }
+
+        final ActionBar actionBar = getSupportActionBar();
+        if (actionBar != null) {
+            actionBar.setDisplayHomeAsUpEnabled(false);
+        }
+
+        updateDrawerNavigation();
+
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(@NonNull final MenuItem item) {
+        if (DEBUG) {
+            Log.d(TAG, "onOptionsItemSelected() called with: item = [" + item + "]");
+        }
+
+        if (item.getItemId() == android.R.id.home) {
+            onHomeButtonPressed();
+            return true;
+        }
+        return super.onOptionsItemSelected(item);
+    }
 
     /*//////////////////////////////////////////////////////////////////////////
     // Init
@@ -337,10 +816,8 @@ public class MainActivity extends AppCompatActivity {
 
         final Fragment fragment = getSupportFragmentManager()
                 .findFragmentById(R.id.fragment_holder);
-
         if (fragment instanceof MainFragment) {
             getSupportActionBar().setDisplayHomeAsUpEnabled(false);
-            final ActionBarDrawerToggle toggle = drawer.getToggle();
             if (toggle != null) {
                 toggle.syncState();
                 toolbarLayoutBinding.toolbar.setNavigationOnClickListener(v -> mainBinding.getRoot()
